+++ src/App.jsx
@@ -3,21 +3,12 @@ import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
 
-// Created an array of objects  
-const todoList = [
-  {
-    id:1,
-    title: "Complete assignment 1"
-  },
-  {
-    id:2,
-    title: "Complete assignment 2"
-  },
-  {
-    id:3,
-    title: "Complete assignment 3"
-  }
-];
+// importing TodoList
+import TodoList from './TodoList';
+// importing TodoForm
+import AddTodoForm from './AddTodoForm';
+
+
 
 function App() {
   const [count, setCount] = useState(0)
@@ -25,11 +16,9 @@ function App() {
   return (
     <>
       <h1>To-Do List</h1>
-      <ul>
-        {todoList.map((todo) => (
-          <li key={todo.id}>{todo.title}</li>
-        ))}
-      </ul>
+      <AddTodoForm/>
+      {/* THis is how you add a comment. Also imported TodoList */}
+      <TodoList />
     </>
   );
 }
